G-2.0-Inventario - Inventario disponible para los empleados

+++ resources/views/admin/inventory/list.blade.php
@@ -25,7 +25,9 @@
                             <th class="text-center">Cantidad</th>
                             <th class="text-right">Costo</th>
                             <th class="text-center">Fecha de Creación</th>
-                            <th class="text-center px-3">Acción</th>
+                            @if (auth()->user()->role == 1)
+                                <th class="text-center px-3">Acción</th>
+                            @endif
                         </tr>
                     </thead>
                     <tbody>
@@ -37,31 +39,33 @@
                             <td class="text-center">{{ $inventory->local }}</td>
                             <td class="text-right">{{ number_format($inventory->cost, 2, ',', '.') }}</td>
                             <td class="text-center">{{ date('d-m-Y', strtotime($inventory->created_at)) }}</td>
-                            <td class="text-center px-3"> 
-                                <div class="d-flex">
-                                    
-                                    <button class="btn btn-sm btn-info me-1" 
-                                        onclick="editInventory(
-                                        {{ $inventory->id }}, 
-                                        {{ $inventory->cost }},
-                                        {{ $inventory->local }},)">
-                                        <i data-feather="edit"></i>                                   
-                                    </button>    
-                                    
-                                    <span class="btn btn-sm btn-danger"
-                                    onclick="deleteElement( {{ $inventory->id }}, 
-                                    '#delete_inventory_', 
-                                    'este Inventario',
-                                    'IMPORTANTE: Si esté Producto del Inventario está añadido en uno de los platos, no podrá ser removido' )"> 
-                                        <i data-feather="trash-2"></i> 
-                                    </span> 
-                        
-                                    <form id="delete_inventory_{{ $inventory->id }}" action="{{ route('inventory.destroy', $inventory->id) }}" method="POST">
-                                        @csrf
-                                        @method('DELETE')                                      
-                                    </form>
-                                </div>
-                            </td>
+                            @if (auth()->user()->role == 1)
+                                <td class="text-center px-3"> 
+                                    <div class="d-flex">
+                                        
+                                        <button class="btn btn-sm btn-info me-1" 
+                                            onclick="editInventory(
+                                            {{ $inventory->id }}, 
+                                            {{ $inventory->cost }},
+                                            {{ $inventory->local }},)">
+                                            <i data-feather="edit"></i>                                   
+                                        </button>    
+                                        
+                                        <span class="btn btn-sm btn-danger"
+                                        onclick="deleteElement( {{ $inventory->id }}, 
+                                        '#delete_inventory_', 
+                                        'este Inventario',
+                                        'IMPORTANTE: Si esté Producto del Inventario está añadido en uno de los platos, no podrá ser removido' )"> 
+                                            <i data-feather="trash-2"></i> 
+                                        </span> 
+                            
+                                        <form id="delete_inventory_{{ $inventory->id }}" action="{{ route('inventory.destroy', $inventory->id) }}" method="POST">
+                                            @csrf
+                                            @method('DELETE')                                      
+                                        </form>
+                                    </div>
+                                </td>
+                            @endif
                         </tr>
                         @endforeach
                     </tbody>
