+++ resources/views/admin/products/list.blade.php
@@ -24,7 +24,9 @@
                             <th class="text-center">Nombre</th>
                             <th class="text-center">Presentación</th>
                             <th class="text-center">Fecha de Creación</th>
-                            <th class="text-center">Acción</th>
+                            @if (auth()->user()->role == 1)
+                                <th class="text-center">Acción</th>
+                            @endif
                         </tr>
                     </thead>
                     <tbody>
@@ -39,41 +41,43 @@
                                 <td class="text-center">{{ $product->quantity }} Unid.</td>
                             @endif
                             <td class="text-center">{{ date('d-m-Y', strtotime($product->created_at)) }}</td>
-                            <td class="text-center"> 
-                                @if ($product->gr != null)
-                                    <button class="btn btn-sm btn-info my-1"
-                                        onclick="editProduct(
-                                        {{ $product->id }}, 
-                                        '{{ $product->name }}',
-                                        'Gramos',
-                                        {{ $product->gr }},
-                                        {{ $product->it_has_flavors }})"> 
-                                        <i data-feather="edit"></i> 
-                                    </button> 
-                                @else
-                                    <button class="btn btn-sm btn-info my-1"
-                                        onclick="editProduct(
-                                        {{ $product->id }}, 
-                                        '{{ $product->name }}',
-                                        'Unidades',
-                                        {{ $product->quantity }},
-                                        {{ $product->it_has_flavors }})"> 
-                                        <i data-feather="edit"></i> 
-                                    </button> 
-                                @endif
+                            @if (auth()->user()->role == 1)
+                                <td class="text-center"> 
+                                    @if ($product->gr != null)
+                                        <button class="btn btn-sm btn-info my-1"
+                                            onclick="editProduct(
+                                            {{ $product->id }}, 
+                                            '{{ $product->name }}',
+                                            'Gramos',
+                                            {{ $product->gr }},
+                                            {{ $product->it_has_flavors }})"> 
+                                            <i data-feather="edit"></i> 
+                                        </button> 
+                                    @else
+                                        <button class="btn btn-sm btn-info my-1"
+                                            onclick="editProduct(
+                                            {{ $product->id }}, 
+                                            '{{ $product->name }}',
+                                            'Unidades',
+                                            {{ $product->quantity }},
+                                            {{ $product->it_has_flavors }})"> 
+                                            <i data-feather="edit"></i> 
+                                        </button> 
+                                    @endif
 
-                                <button class="btn btn-sm btn-danger"
-                                    onclick="deleteElement( {{ $product->id }}, 
-                                    '#delete_product_', 
-                                    'este Producto',
-                                    'IMPORTANTE: Si esté Producto está añadido en el inventario, no podrá ser removido' )"> 
-                                    <i data-feather="trash-2"></i> 
-                                </button>
-                                <form id="delete_product_{{ $product->id }}" action="{{ route('products.destroy', $product->id) }}" method="POST">
-                                    @csrf
-                                    @method('DELETE')                                      
-                                </form>
-                            </td>
+                                    <button class="btn btn-sm btn-danger"
+                                        onclick="deleteElement( {{ $product->id }}, 
+                                        '#delete_product_', 
+                                        'este Producto',
+                                        'IMPORTANTE: Si esté Producto está añadido en el inventario, no podrá ser removido' )"> 
+                                        <i data-feather="trash-2"></i> 
+                                    </button>
+                                    <form id="delete_product_{{ $product->id }}" action="{{ route('products.destroy', $product->id) }}" method="POST">
+                                        @csrf
+                                        @method('DELETE')                                      
+                                    </form>
+                                </td>
+                            @endif
                         </tr>
                         @endforeach
                     </tbody>
